requirements.txt
# pip install -r requirements.txt

# pip==20.1.1
# base ----------------------------------------
# cv2 # >=4.5.5
Cython
matplotlib>=3.2.2
numpy>=1.18.5 # 1.25.0
# my comment opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
scipy>=1.4.1 # 1.8.0
tensorboard>=2.2
torch>=1.7.0  # 1.10.0+cpu
torchvision>=0.8.1 #  0.10.0
tqdm>=4.41.0

# logging -------------------------------------
# wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas

# export --------------------------------------
# coremltools>=4.1
# onnx>=1.8.1
# scikit-learn==0.19.2  # for coreml quantization

# extras --------------------------------------
thop  # FLOPS computation
# me=y comment pycocotools>=2.0  # COCO mAP
